+++ src/app/cart/components/cart-item/cart-item.component.ts
@@ -1,4 +1,5 @@
 import { Component, Input, Output, EventEmitter } from '@angular/core';
+import { ProductList } from './../../../core/interfaces/product-list';
 
 @Component({
   selector: 'app-cart-item',