+++ src/app/cart/components/cart-list/cart-list.component.ts
@@ -1,5 +1,6 @@
 import { Component, ChangeDetectionStrategy, DoCheck } from '@angular/core';
 import { CartService } from 'src/app/cart/services/cart.service';
+import { ProductList } from './../../../core/interfaces/product-list';
 
 @Component({
   selector: 'app-cart-list',